fix: to_prev_start() multi-level consecutive folds
Because (n)vim doesn't offer a function to detect consecutive folds, the
previous implementation used `zk` in combination with counting indents
to detect the most recently opened fold. However this wouldn't work in
the case where `zk` didn't navigate immediately to a consecutive fold.

This new implementation should be more robust - it always tries to go
back at least as far as it should, and then runs `zj` up until the point
that the original line would be hit.